+++ data/archive.json
@@ -1,6 +1,6 @@
 {
   "archive": {
-    "banList": ["portfoliov2", "virtulib-project", "calgaryhacks2022"],
+    "banList": [],
     "projectList": []
   }
 }
