+++ data/sections.json
@@ -1,3 +1,4 @@
 {
-  "sections": ["Intro", "About", "Experience", "Projects", "Archive", "Contact"]
+  "sections": ["About", "Experience", "Projects", "Contact"],
+  "toolbarWidth": "10rem"
 }
