+++ section_components/index.js
@@ -2,7 +2,6 @@ import Intro from "./Intro";
 import About from "./About";
 import Experience from "./Experience";
 import Projects from "./Projects";
-import Archive from "./Archive";
 import Contact from "./Contact";
 
-export { Intro, About, Experience, Projects, Archive, Contact };
+export { Intro, About, Experience, Projects, Contact };
